+++ packages/liveblocks-react/src/ThreadDB.ts
@@ -1,5 +1,7 @@
 import type {
   BaseMetadata,
+  SubscriptionData,
+  SubscriptionKey,
   ThreadData,
   ThreadDataWithDeleteInfo,
   ThreadDeleteInfo,
@@ -178,12 +180,15 @@ export class ThreadDB<M extends BaseMetadata> {
    *   'desc' means by updatedAt DESC
    *
    * Will never return deleted threads in the result.
+   *
+   * Subscriptions are needed to filter threads based on the user's subscriptions.
    */
   public findMany(
     // TODO: Implement caching here
     roomId: string | undefined,
     query: ThreadsQuery<M> | undefined,
-    direction: "asc" | "desc"
+    direction: "asc" | "desc",
+    subscriptions: Record<SubscriptionKey, SubscriptionData> | undefined
   ): ThreadData<M>[] {
     const index = direction === "desc" ? this.#desc : this.#asc;
     const crit: ((thread: ThreadData<M>) => boolean)[] = [];
@@ -191,7 +196,7 @@ export class ThreadDB<M extends BaseMetadata> {
       crit.push((t) => t.roomId === roomId);
     }
     if (query !== undefined) {
-      crit.push(makeThreadsFilter(query));
+      crit.push(makeThreadsFilter(query, subscriptions));
     }
     return Array.from(index.filter((t) => crit.every((pred) => pred(t))));
   }