+++ CHANGELOG.md
@@ -1,4 +1,17 @@
-## vNEXT (not yet published)
+## v3.9.0 (not yet published)
+
+### `@liveblocks/react`
+
+- Add query filters `subscribed` on the `useThreads` hook.
+
+### `@liveblocks/client`
+
+- Add query filter `subscribed` on the `room.getThreads` method.
+
+### `@liveblocks/node`
+
+- Remove all schema validation related client methods that should no longer be
+  used. Schema validation was sunsetted on May 1st, 2025.
 
 ## v3.8.1